Refactor some methods to ORders and Products repository to make user more consistent

+++ MundoBalloonApi.GraphQL/Users/UserCartMutations.AddToCart.cs
@@ -7,7 +7,7 @@ namespace MundoBalloonApi.graphql.Users;
 public partial class UserCartMutations
 {
     [Authorize]
-    public Task<UserCartProduct> AddToCart(string sku, double quantity, double price, int productVariantId,
+    public Task<UserCartProduct> AddToCart(string sku, decimal quantity, decimal price, int productVariantId,
         [Service] IUsersService usersService, [GlobalState("currentUser")] CurrentUser currentUser,
         CancellationToken cancellationToken)
     {
