+++ MundoBalloonApi.GraphQL/Users/UserEventMutations.AddToEventCart.cs
@@ -8,7 +8,7 @@ public partial class UserEventMutations
 {
     [Authorize]
     public Task<EventCartDetail> AddToEventCart(
-        int productVariantId, int userEventId, double quantity,
+        int productVariantId, int userEventId, decimal quantity,
         [Service] IUsersService usersService,
         CancellationToken cancellationToken)
     {
