+++ MundoBalloonApi.infrastructure/Data/Models/EventCartDetail.cs
@@ -4,7 +4,7 @@ public class EventCartDetail : BaseEntity
 {
     public int EventCartDetailId { get; init; }
     public int ProductVariantId { get; init; }
-    public double Quantity { get; init; }
+    public decimal Quantity { get; init; }
     public int? UserEventId { get; init; }
 
     public ProductVariant? ProductVariant { get; set; }
