+++ MundoBalloonApi.infrastructure/Data/Models/UserCartProduct.cs
@@ -4,8 +4,8 @@ public class UserCartProduct : BaseEntity
 {
     public string? UserId { get; init; }
     public string? Sku { get; init; }
-    public double Quantity { get; set; }
-    public double Price { get; init; }
+    public decimal Quantity { get; set; }
+    public decimal Price { get; init; }
     public int? ProductVariantId { get; init; }
 
     public ProductVariant? ProductVariant { get; set; }
